Add entity descriptions + hitbox viewer
Adding a high-level overview of how entities work in the game in a new
"Internal_Structures" directory, along with what I know so far of the
layout.

Also, adding v4.5 of the hitbox viewer script. This has been confirmed
to work with BizHawk 2.9.1 onwards (although BizHawk 2.10 has a bug
which causes the floating text to be displayed at incorrect positions).